Created a transition_state_machine.py and corresponding transition_state.py.
It was necessary to create a state machine since the feedback action could not handle the outcomes set for the specified transition. To create a structured overview the transition state machine will execute the transition and if this is correct it will move into the new state. This structure has been added to the healthy state machine.